requirements.txt
# ─────────────────────────────────────────────
# MTA Enterprise Workforce System — Requirements
# ─────────────────────────────────────────────

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
mangum==0.17.0          # ASGI adapter for AWS Lambda / Vercel

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9  # PostgreSQL driver
alembic==1.13.1         # DB migrations

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Scheduling
apscheduler==3.10.4

# BigQuery
google-cloud-bigquery==3.22.0
google-auth==2.29.0

# HTTP / Validation
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
tabulate==0.9.0

# Dev / Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
